+++ src/imio/helpers/xhtml.py
@@ -41,3 +41,58 @@ def removeBlanks(xhtmlContent):
             el.getparent().remove(el)
     # only return children of the <special_tag>
     return ''.join([lxml.html.tostring(x, encoding='utf-8', pretty_print=True, method='xml') for x in tree.iterchildren()])
+
+
+def addClassToLastChildren(xhtmlContent,
+                           className='paraKeepWithNext',
+                           tags=('p', 'ul', 'li', ),
+                           numberOfChars=60):
+    '''This method will add a class attribute adding p_className to the last tags of given p_xhtmlContent.
+       It only consider given p_tags and will define the class on last tags until it contains given
+       p_numberOfChars number of characters.
+    '''
+    if not xhtmlContent or not xhtmlContent.strip():
+        return xhtmlContent
+
+    # surround xhtmlContent with a special tag so we are sure that tree is always
+    # a list of children of this special tag
+    tree = lxml.html.fromstring(unicode("<special_tag>%s</special_tag>" % xhtmlContent.strip(), 'utf-8'))
+    children = tree.getchildren()
+    if not children:
+        return xhtmlContent
+
+    def adaptTree(children, managedNumberOfChars=0):
+        """
+          Recursive method that walk the children and subchildren and adapt what necessary.
+        """
+        # apply style on last element until we reached necessary numberOfChars or we encounter
+        # a tag not in p_tags or we do not have a tag...
+        i = 1
+        stillNeedToAdaptPreviousChild = True
+        numberOfChildren = len(children)
+        while stillNeedToAdaptPreviousChild and i <= numberOfChildren:
+            child = children[-i]
+            if not child.tag in tags:
+                stillNeedToAdaptPreviousChild = False
+            else:
+                subchildren = child.getchildren()
+                if subchildren:
+                    # recursion is done here
+                    managedNumberOfChars = adaptTree(subchildren, managedNumberOfChars=managedNumberOfChars)
+                else:
+                    # check if tag did not already have a class attribute
+                    # in this case, we append className to existing classes
+                    if 'class' in child.attrib:
+                        child.attrib['class'] = '{0} {1}'.format(className, child.attrib['class'])
+                    else:
+                        child.attrib['class'] = className
+                    managedNumberOfChars += child.text and len(child.text) or 0
+                if managedNumberOfChars >= numberOfChars:
+                    stillNeedToAdaptPreviousChild = False
+                i = i + 1
+        return managedNumberOfChars
+
+    # call recursive method 'adaptTree' that whill adapt children and subchildren
+    adaptTree(children)
+
+    return ''.join([lxml.html.tostring(x, encoding='utf-8', pretty_print=True, method='xml') for x in tree.iterchildren()])